Correct attribute assignments for ReversibleRxn()
It was calling rxn.[x] items that didn't exist, and it also failed to
call the self.parse() function used in ElementaryRxn().
Now, it parses the input file and assigns some attributes to new
attributes (not sure if this is necessary or if we should just clean it
up to use, e.g., self.r_stoich rather than re-assigning this to self.r
after parsing).
Also added an assertion that user-entered T (temperature) is numeric.